#3607: re-use code
